Fix: path 이름 수정, 회원탈퇴 패스 추가
